No need to have a global variable.
I think having a global variable is not a really good practice, so I came up with a workaround to determine whether or not a comma must be printed into the stdout. It just checks char by char the `stack` values and compares it with the number range max value.